move patch rotation stuff into their own files
