+++ src/Makefile
@@ -552,6 +552,7 @@ OBJS:=$(i_main_o) \
 		$(OBJDIR)/r_sky.o    \
 		$(OBJDIR)/r_splats.o \
 		$(OBJDIR)/r_patch.o  \
+		$(OBJDIR)/r_patchrotation.o  \
 		$(OBJDIR)/r_things.o \
 		$(OBJDIR)/r_portal.o \
 		$(OBJDIR)/screen.o   \
